Search refactor.
