+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "spaczz"
-version = "0.3.2"
+version = "0.4.0"
 description = "Adds fuzzy matching and additional regex matching support to spaCy."
 license = "MIT"
 readme = "README.md"
