--- src/spaczz/fuzz/__init__.py
@@ -1,5 +1,0 @@
-"""Module for fuzzy matching components."""
-from .fuzzysearcher import FuzzySearcher
-
-
-__all__ = ["FuzzySearcher"]
