+++ src/spaczz/regex/__init__.py
@@ -1,6 +1,5 @@
 """Module for regex components."""
 from .regexconfig import RegexConfig
-from .regexsearcher import RegexSearcher
 
 
-__all__ = ["RegexConfig", "RegexSearcher"]
+__all__ = ["RegexConfig"]
