+++ src/spaczz/search/__init__.py
@@ -1,0 +1,7 @@
+"""Module for search components."""
+from ._phrasesearcher import _PhraseSearcher
+from .fuzzysearcher import FuzzySearcher
+from .regexsearcher import RegexSearcher
+from .similaritysearcher import SimilaritySearcher
+
+__all__ = ["_PhraseSearcher", "FuzzySearcher", "RegexSearcher", "SimilaritySearcher"]
