+++ src/spaczz/search/regexsearcher.py
@@ -1,35 +1,47 @@
-"""Module for RegexSearcher. Does multi-token regex matching in spaCy Docs."""
+"""Module for RegexSearcher: multi-token regex matching in spaCy `Doc` objects."""
 from typing import List, Tuple, Union
 
 from spacy.tokens import Doc, Span
+from spacy.vocab import Vocab
 
-from .regexconfig import RegexConfig
 from ..process import map_chars_to_tokens
+from ..regex import RegexConfig
 
 
 class RegexSearcher:
-    """Class for multi-token regex matching in spacy Docs.
+    """Class for multi-token regex matching in spacy `Doc` objects.
 
     Regex matching is done on the character level and then
     mapped back to tokens.
 
     Attributes:
+        vocab (Vocab): The shared vocabulary.
+            Included for consistency and potential future-state.
         _config (RegexConfig): The regex config used with the
             regex searcher.
     """
 
-    def __init__(self, config: Union[str, RegexConfig] = "default") -> None:
+    def __init__(
+        self, vocab: Vocab, config: Union[str, RegexConfig] = "default"
+    ) -> None:
         """Initializes the regex searcher with the given config.
 
         Args:
+            vocab: A spaCy `Vocab` object.
+                Purely for consistency between spaCy
+                and spaczz matcher APIs for now.
+                spaczz matchers are mostly pure-Python
+                currently and do not share vocabulary
+                with spaCy pipelines.
             config: Provides the class with predefind regex patterns.
                 Uses the default config if "default", an empty config if "empty",
                 or a custom config by passing a RegexConfig object.
                 Default is "default".
 
         Raises:
-            TypeError: If config is not a RegexConfig object.
+            TypeError: If config is not a `RegexConfig` object.
         """
+        self.vocab = vocab
         if config == "default":
             self._config = RegexConfig(empty=False)
         elif config == "empty":
@@ -47,7 +59,7 @@ class RegexSearcher:
                 )
 
     def match(
-        self, doc: Doc, regex_str: str, partial: bool = True, predef: bool = False,
+        self, doc: Doc, query: str, partial: bool = True, predef: bool = False,
     ) -> List[Tuple[int, int, Tuple[int, int, int]]]:
         """Returns all the regex matches within doc.
 
@@ -60,8 +72,8 @@ class RegexSearcher:
         To utilize regex flags, use inline flags.
 
         Args:
-            doc: Doc object to search over.
-            regex_str: A string to be compiled to regex,
+            doc: `Doc` object to search over.
+            query: A string to be compiled to regex,
                 or the key name of a predefined regex pattern.
             partial: Whether partial matches should be extended
                 to existing span boundaries in doc or not, i.e.
@@ -91,21 +103,21 @@ class RegexSearcher:
             A list of span start index, end index, fuzzy change count tuples.
 
         Raises:
-            TypeError: If regex_str is not a string.
+            TypeError: If query is not a string.
 
         Example:
             >>> import spacy
-            >>> from spaczz.regex import RegexSearcher
+            >>> from spaczz.search import RegexSearcher
             >>> nlp = spacy.blank("en")
             >>> searcher = RegexSearcher()
-            >>> doc = nlp.make_doc("My phone number is (555) 555-5555.")
+            >>> doc = nlp("My phone number is (555) 555-5555.")
             >>> searcher.match(doc, "phones", predef=True)
             [(4, 10, (0, 0, 0))]
         """
-        if isinstance(regex_str, str):
-            compiled_regex = self._config.parse_regex(regex_str, predef)
+        if isinstance(query, str):
+            compiled_regex = self._config.parse_regex(query, predef)
         else:
-            raise TypeError(f"regex_str must be a str, not {type(regex_str)}.")
+            raise TypeError(f"query must be a str, not {type(query)}.")
         matches = []
         chars_to_tokens = map_chars_to_tokens(doc)
         for match in compiled_regex.finditer(doc.text):
