--- src/spaczz/similarity/__init__.py
@@ -1,5 +1,0 @@
-"""Module for similarity matching components."""
-from .similaritysearcher import SimilaritySearcher
-
-
-__all__ = ["SimilaritySearcher"]
