+++ tests/conftest.py
@@ -5,12 +5,12 @@ from spacy.language import Language
 
 
 @pytest.fixture
-def nlp() -> Language:
-    """Empty spacy English language pipeline."""
-    return spacy.blank("en")
-
-
-@pytest.fixture
 def model() -> Language:
     """Medium English Core spaCy model."""
     return spacy.load("en_core_web_md")
+
+
+@pytest.fixture
+def nlp() -> Language:
+    """Empty spaCy English language pipeline."""
+    return spacy.blank("en")
